Create Day_04 (Array Problem)
I learned some array-type problems in C++ that I solved in 1h which was really disappointing for me because this problem has taken a huge time from me. I did not expect it while I was think that was easy for me but it was a little bit but finally solved it. I need to more practice about this problem.