Create Day_06 ( String Problem)
I learned string-type problems in C++ and tried to apply them in CodeChef. Finally, I am completed today's challenge. I hope as soon as possible I will jump to an advanced-level problem which is the upcoming programming contest. I do not know how can do it but find out the path. That's all for today.