Create Day_13 (Previous problem Solved)
Today I was so busy at work. I did not practice the new problem in CodeChef but I tried to something solve the previous problem which is completed by the challenge. Now I solved the previous topic in the basic problem for today. I do have not any energy to learn something new.