Create Day_14 (Math Problem)
Today I solved the intermediate math problem in CodeChef. It was not so complicated but I did the complication while I was solving it. After some hours I got the point then easily solved my idea but when I understood the problem and solved it then I thought about myself it is known to all but I did not remind the working time.